+++ Shuhari.Framework.Wpf.GuiTests/TestDialog.xaml.cs
@@ -1,0 +1,15 @@
+﻿using Shuhari.Framework.Wpf.Controls;
+
+namespace Shuhari.Framework.Wpf.GuiTests
+{
+    /// <summary>
+    /// Interaction logic for TestDialog.xaml
+    /// </summary>
+    public partial class TestDialog : Dialog
+    {
+        public TestDialog()
+        {
+            InitializeComponent();
+        }
+    }
+}