+++ Shuhari.Framework.Wpf/Controls/ElementRole.cs
@@ -1,0 +1,23 @@
+﻿namespace Shuhari.Framework.Wpf.Controls
+{
+    /// <summary>
+    /// Special role for element
+    /// </summary>
+    public enum ElementRole
+    {
+        /// <summary>
+        /// None
+        /// </summary>
+        None = 0,
+
+        /// <summary>
+        /// Ok button. currently not used
+        /// </summary>
+        Ok = 1,
+
+        /// <summary>
+        /// Click to cancel dialog with no change
+        /// </summary>
+        Cancel = 2,
+    }
+}